+++ example/lib/features/team/presentation/controllers/members_controller.dart
@@ -1,4 +1,4 @@
-import 'package:base_core/base_core.dart';
+import 'package:clean_signals/clean_signals.dart';
 import 'package:signals/signals.dart';
 
 import '../../domain/entities/member.dart';