+++ test/controller_test.dart
@@ -1,6 +1,6 @@
 import 'dart:async';
 
-import 'package:base_core/base_core.dart';
+import 'package:clean_signals/clean_signals.dart';
 import 'package:signals/signals.dart';
 import 'package:test/test.dart';